+++ frontend/components/organisms/GroupJoinBtn.vue
@@ -1,0 +1,56 @@
+<template>
+  <div>
+    <v-btn
+      color="#1CA1F1"
+      rounded
+      outlined
+      @click.prevent="() => requestGroupJoin(group)"
+    >
+      グループに参加する
+    </v-btn>
+
+    <v-snackbar
+      v-model="join"
+      :color="resultJoinType"
+      top
+      vertical
+      :timeout="2500"
+    >
+      {{ resultJoinMessage }}
+      <v-btn dark text @click="join = false">CLOSE</v-btn>
+    </v-snackbar>
+  </div>
+</template>
+
+<script>
+export default {
+  props: {
+    group: {
+      type: Object,
+      required: true
+    }
+  },
+
+  data: () => ({
+    join: false,
+    resultJoinType: null,
+    resultJoinMessage: null
+  }),
+
+  methods: {
+    async requestGroupJoin(group) {
+      await this.$axios
+        .$post(`/api/groups/${group.id}/join`)
+        .then(() => {
+          this.resultJoinMessage = `「${group.name}」グループに参加しました`
+          this.resultJoinType = 'info'
+        })
+        .catch(() => {
+          this.resultJoinMessage = `グループに参加失敗しました`
+          this.resultJoinType = 'error'
+        })
+      this.join = true
+    }
+  }
+}
+</script>
